Added sleep so i wouldnt miss files added
its kind of hacky, but the idea here is that the first file trigegers an S3 event, the subsequent ones do not as they are prevented from running by the `.processing` file, so what we do is just wait for 20 seconds before getting the list of files. That way, all files should finish uploading by the time the script runs. I don't lik eit, but i am not sure how else to do it, maybe read files, wait 5 seconds, read files again, if files list is the same, continue, otherwise wait 5 more seconds?